Remove InfoDialog and code placed back in Audit
