--- src/components/InfoDialog.vue
@@ -1,104 +1,0 @@
-<template>
-  <v-dialog v-model="viewDialogChild" max-width="600">
-    <div v-if="dialogReq !== undefined">
-      <v-card>
-        <v-btn icon flat style="float:right" @click="viewDialogChild = false">
-          <v-icon>close</v-icon>
-        </v-btn>
-        <v-card-title>{{ dialogReq["title"] }}</v-card-title>
-        <v-card-text v-if="'url' in dialogReq">
-          <a target="_blank" :href="dialogReq['url']">Link to the real, up to date requirements</a>
-        </v-card-text>
-        <v-card-text v-if="'desc' in dialogReq">
-          {{ dialogReq["desc"] }}
-        </v-card-text>
-        <v-card-text>
-          <div
-            class="percentage-bar p-block"
-            :style="percentage(dialogReq)"
-          >
-            {{ dialogReq["percent_fulfilled"] }}% fulfilled
-          </div>
-        </v-card-text>
-        <v-card-text v-if="'req' in dialogReq">
-          {{ dialogReq["req"] }}
-        </v-card-text>
-        <v-card-text>
-          <b>Satisfying courses:</b>
-          <div v-for="course in dialogReq['sat_courses']" :key="course">
-            {{ course }}
-          </div>
-        </v-card-text>
-        <v-card-actions>
-          <v-spacer />
-          <v-btn
-            v-if="'title-no-degree' in dialogReq"
-            color="error"
-            @click="deleteReq(dialogReq); viewDialogChild = false;"
-          >
-            <v-icon>delete</v-icon>
-            Remove Requirement
-          </v-btn>
-        </v-card-actions>
-      </v-card>
-    </div>
-  </v-dialog>
-</template>
-
-<script>
-export default {
-  name: 'InfoDialog',
-  props: [
-    'selectedSubjects',
-    'viewDialog',
-    'dialogReq'
-  ],
-  data: function () {
-    return {
-      viewDialogChild: false
-    };
-  },
-  watch: {
-    viewDialogChild (newVal) {
-      this.$emit('update:view-dialog', newVal);
-      if (newVal === false) {
-        this.$emit('update:dialog-req', undefined);
-      }
-    },
-    viewDialog (newVal) {
-      this.viewDialogChild = newVal;
-    }
-  },
-  methods: {
-    percentage: function (req) {
-      const pfulfilled = req.percent_fulfilled;
-      const pcolor = req.fulfilled
-        ? '#00b300'
-        : req.percent_fulfilled > 15
-          ? '#efce15'
-          : '#ef8214';
-      return `--percent: ${pfulfilled}%; --bar-color: ${pcolor}; --bg-color: #fff`;
-    },
-    deleteReq: function (req) {
-      const reqName = req['list-id'].substring(0, req['list-id'].indexOf('.reql'));
-      this.$store.commit('removeReq', reqName);
-    }
-  }
-};
-</script>
-<style scoped>
-.percentage-bar {
-  background: linear-gradient(
-    90deg,
-    var(--bar-color) var(--percent),
-    var(--bg-color) var(--percent)
-  );
-}
-.p-block {
-  height: 30px;
-  border: 1px solid gray;
-  padding-left: 5px;
-  padding-top: 5px;
-  padding-bottom: 5px;
-}
-</style>
